requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26
sqlite-utils>=3.37
tabulate>=0.9.0
requests>=2.32.3
tqdm>=4.66.4
openai>=1.40.0
tavily-python>=0.5.0
kaggle>=1.6.14